move SeedMessenger to its own file
